Update .idea/.gitignore: add Book-env to ignores
This commit updates the .idea/.gitignore file by adding "Book-env" to the list of ignored items. The previous version of the file included a line with "/Book-env" that was accidentally prefixed with a slash, which would only ignore a subdirectory named "Book-env" in the root directory. This commit removes the slash to ignore the "Book-env" file or directory at any location.